+++ src/ResourceManager/DataLakeAnalytics/Commands.DataLakeAnalytics/Commands/TestAzureRmDataLakeAnalyticsAccount.cs
@@ -18,7 +18,7 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.DataLakeAnalytics
 {
-    [Cmdlet(VerbsDiagnostic.Test, "AzureRmDataLakeAnalyticsAccount"), OutputType(typeof(bool))]
+    [Cmdlet("Test", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataLakeAnalyticsAccount"), OutputType(typeof(bool))]
     [Alias("Test-AdlAnalyticsAccount")]
     public class TestAzureDataLakeAnalyticsAccount : DataLakeAnalyticsCmdletBase
     {