+++ src/ResourceManager/KeyVault/Commands.KeyVault/Commands/UpdateAzureKeyVaultKey.cs
@@ -23,9 +23,7 @@ namespace Microsoft.Azure.Commands.KeyVault
     /// Update attribute of a key vault key.
     /// </summary>
     [Alias("Set-AzureKeyVaultKey", "Set-AzureKeyVaultKeyAttribute")]
-    [Cmdlet(VerbsData.Update, "AzureKeyVaultKey",
-        SupportsShouldProcess = true, 
-        DefaultParameterSetName = DefaultParameterSet)]
+    [Cmdlet("Update", ResourceManager.Common.AzureRMConstants.AzurePrefix + "KeyVaultKey",SupportsShouldProcess = true,DefaultParameterSetName = DefaultParameterSet)]
     [OutputType(typeof(PSKeyVaultKey))]
     public class UpdateAzureKeyVaultKey : KeyVaultCmdletBase
     {