+++ src/ResourceManager/Network/Commands.Network/Models/PSRetentionPolicyParameters.cs
@@ -12,6 +12,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.WindowsAzure.Commands.Common.Attributes;
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
@@ -20,9 +21,11 @@ namespace Microsoft.Azure.Commands.Network.Models
     public class PSRetentionPolicyParameters
     {
         [JsonProperty(Order = 2)]
+        [Ps1Xml(Target = ViewControl.Table)]
         public int? Days { get; set; }
 
         [JsonProperty(Order = 2)]
+        [Ps1Xml(Target = ViewControl.Table)]
         public bool? Enabled { get; set; }
     }
 }