+++ src/ResourceManager/AnalysisServices/Commands.AnalysisServices/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 0.6.14
+ * Fixed issue with default resource groups not being set.
+ * Updated common runtime assemblies
+
+## Version 0.6.13
+* Fixed issue with default resource groups not being set.
+
 ## Version 0.6.12
 * Updated to the latest version of the Azure ClientRuntime.