+++ src/ResourceManager/DevSpaces/Commands.DevSpaces/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 0.0.5
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 0.0.4
+* Fixed issue with default resource groups not being set.
+
 ## Version 0.0.3
 * Updated to the latest version of the Azure ClientRuntime.