+++ src/ResourceManager/MachineLearningCompute/Commands.MachineLearningCompute/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 0.4.8
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 0.4.7
+* Fixed issue with default resource groups not being set.
+
 ## Version 0.4.6
 * Updated to the latest version of the Azure ClientRuntime.